refactor: centralize firebase initialization
