+++ script-ads.js
@@ -1,11 +1,9 @@
 import { encryptString, decryptString } from './crypto.js';
 import logger from './logger.js';
+import { db, auth } from './src/firebase.js';
+import { onAuthStateChanged } from 'https://www.gstatic.com/firebasejs/9.22.2/firebase-auth.js';
 
-if (!firebase.apps.length) {
-  firebase.initializeApp(firebaseConfig);
-}
-const db = firebase.firestore();
-firebase.auth().onAuthStateChanged((user) => {
+onAuthStateChanged(auth, (user) => {
   if (!user) {
     window.location.href = 'index.html?login=1';
   }
