--- scripts/firebase-init.js
@@ -1,24 +1,0 @@
-import { initializeApp, getApps } from 'https://www.gstatic.com/firebasejs/9.22.2/firebase-app.js';
-import {
-  getFirestore,
-  collection,
-  doc,
-  setDoc
-} from 'https://www.gstatic.com/firebasejs/9.22.2/firebase-firestore.js';
-
-// Configuração Firebase
-const firebaseConfig = {
-  apiKey: "AIzaSyC78l9b2DTNj64y_0fbRKofNupO6NHDmeo",
-  authDomain: "matheus-35023.firebaseapp.com",
-  projectId: "matheus-35023",
-  storageBucket: "matheus-35023.appspot.com",
-  messagingSenderId: "1011113149395",
-  appId: "1:1011113149395:web:c1f449e0e974ca8ecb2526"
-};
-
-// Inicializa Firebase
-const app = getApps().length ? getApps()[0] : initializeApp(firebaseConfig);
-const db = getFirestore(app);
-
-// 🔁 Exporte tudo que o background precisa:
-export { db, collection, doc, setDoc };
